pyproject.toml
[project]
name = "fitkeeper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.84.0",
    "fastapi>=0.135.1",
    "numpy>=2.4.3",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.3.0",
    "supabase>=2.28.0",
    "uvicorn>=0.41.0",
]
